refactor: be explicit about coupling to errors
