+++ src/clients/index.ts
@@ -1,4 +1,5 @@
-import type { error, GitHub, GitHubClient, Inputter } from "./types";
+import type { error } from "../errors";
+import type { GitHub, GitHubClient, Inputter } from "./types";
 
 import errors from "../errors";
 import Client from "./client";
