+++ src/helpers/filters.ts
@@ -1,4 +1,5 @@
-import type { error, GitHub, GitHubClient } from "./types";
+import type { error } from "../errors";
+import type { GitHub, GitHubClient } from "./types";
 
 import filters from "../filters";
 import { EVENT_PULL_REQUEST, EVENT_PUSH } from "./constants";
