+++ src/helpers/types.d.ts
@@ -1,8 +1,7 @@
 import type { GitHubClient } from "../clients";
-import type { error, GitHub } from "../types";
+import type { GitHub } from "../types";
 
 export type {
-  error,
   GitHub,
   GitHubClient,
 };
