+++ src/outputs/names.ts
@@ -1,4 +1,4 @@
-import type { error } from "./types";
+import type { error } from "../errors";
 
 const enum OutputName {
   DID_OPTIMIZE = "DID_OPTIMIZE",
