+++ src/parsers/builder.ts
@@ -1,4 +1,5 @@
-import type { error, ParseFn, SafeParseFn } from "./types";
+import type { error } from "../errors";
+import type { ParseFn, SafeParseFn } from "./types";
 
 import errors from "../errors";
