+++ src/parsers/types.d.ts
@@ -1,11 +1,10 @@
-import type { error } from "../types";
+import type { error } from "../errors";
 
 type ParseFn<OutType> = (raw: string) => OutType;
 
 type SafeParseFn<OutType> = (raw: string) => [OutType, error];
 
 export type {
-  error,
   ParseFn,
   SafeParseFn,
 };
