+++ src/svgo/project.ts
@@ -1,4 +1,5 @@
-import type { error, SVGOptimizer } from "./types";
+import type { error } from "../errors";
+import type { SVGOptimizer } from "./types";
 
 import importCwd from "import-cwd";
