+++ src/svgo/types.d.ts
@@ -1,4 +1,4 @@
-import type { error } from "../types";
+import type { error } from "../errors";
 
 type SupportedSvgoVersions = "1" | "2" | "project";
 
@@ -7,7 +7,6 @@ interface SVGOptimizer {
 }
 
 export type {
-  error,
   SupportedSvgoVersions,
   SVGOptimizer,
 };
