+++ src/svgo/v1/index.ts
@@ -1,4 +1,5 @@
-import type { error, SVGOptimizer, SVGOptions } from "./types";
+import type { error } from "../../errors";
+import type { SVGOptimizer, SVGOptions } from "./types";
 
 import svgo from "svgo-v1";
