+++ src/svgo/v1/types.d.ts
@@ -1,4 +1,5 @@
-import type { error, SVGOptimizer } from "../types";
+import type { error } from "../../errors";
+import type { SVGOptimizer } from "../types";
 import type svgo from "svgo-v1";
 
 type SVGOptions = svgo.Options;
