+++ src/svgo/v1/wrapper.ts
@@ -1,4 +1,5 @@
-import type { error, SVGOptions } from "./types";
+import type { error } from "../../errors";
+import type { SVGOptions } from "./types";
 import type svgo from "svgo-v1";
 
 import errors from "../../errors";
